+++ packages/colorpickers/src/elements/Colorpicker/ColorWell.spec.tsx
