+++ packages/colorpickers/src/styled/Colorpicker/StyledColorPicker.ts
