+++ packages/colorpickers/src/styled/Colorpicker/StyledColorWell.spec.tsx
