+++ packages/colorpickers/src/styled/Colorpicker/StyledColorWell.ts
