+++ packages/colorpickers/src/styled/Colorpicker/StyledColorWellGradient.ts
