+++ packages/colorpickers/src/styled/Colorpicker/StyledColorWellThumb.ts
