+++ packages/colorpickers/src/styled/Colorpicker/StyledColorWellWrapper.ts
