+++ packages/colorpickers/src/styled/Colorpicker/StyledInput.ts
