+++ packages/colorpickers/src/styled/Colorpicker/StyledInputGroup.ts
