+++ packages/colorpickers/src/styled/Colorpicker/StyledLabel.ts
