+++ packages/colorpickers/src/styled/Colorpicker/StyledPreview.ts
