+++ packages/colorpickers/src/styled/Colorpicker/StyledSliderGroup.ts
