+++ packages/colorpickers/src/styled/Colorpicker/StyledSliders.spec.tsx
