+++ packages/colorpickers/src/styled/Colorpicker/StyledSliders.ts
