+++ packages/colorpickers/src/styled/ColorpickerDialog/StyledTooltipBody.ts
