+++ packages/colorpickers/src/styled/index.ts
@@ -5,23 +5,23 @@
  * found at http://www.apache.org/licenses/LICENSE-2.0.
  */
 
-export * from './ColorPicker/StyledColorPicker';
-export * from './ColorPicker/StyledHue';
-export * from './ColorPicker/StyledAlpha';
-export * from './ColorPicker/StyledPreview';
-export * from './ColorPicker/StyledColorWell';
-export * from './ColorPicker/StyledColorWellGradient';
-export * from './ColorPicker/StyledColorWellWrapper';
-export * from './ColorPicker/StyledColorWellThumb';
-export * from './ColorPicker/StyledSliderGroup';
-export * from './ColorPicker/StyledHexField';
-export * from './ColorPicker/StyledLabel';
-export * from './ColorPicker/StyledInput';
-export * from './ColorPicker/StyledInputGroup';
-export * from './ColorPicker/StyledRGBAField';
-export * from './ColorPicker/StyledSliders';
+export * from './Colorpicker/StyledColorPicker';
+export * from './Colorpicker/StyledHueRange';
+export * from './Colorpicker/StyledAlphaRange';
+export * from './Colorpicker/StyledPreview';
+export * from './Colorpicker/StyledColorWell';
+export * from './Colorpicker/StyledColorWellGradient';
+export * from './Colorpicker/StyledColorWellWrapper';
+export * from './Colorpicker/StyledColorWellThumb';
+export * from './Colorpicker/StyledSliderGroup';
+export * from './Colorpicker/StyledHexField';
+export * from './Colorpicker/StyledLabel';
+export * from './Colorpicker/StyledInput';
+export * from './Colorpicker/StyledInputGroup';
+export * from './Colorpicker/StyledRGBAField';
+export * from './Colorpicker/StyledSliders';
 
-export * from './ColorDialog/StyledDialogPreview';
-export * from './ColorDialog/StyledButton';
-export * from './ColorDialog/StyledTooltipModal';
-export * from './ColorDialog/StyledTooltipBody';
+export * from './ColorpickerDialog/StyledButton';
+export * from './ColorpickerDialog/StyledButtonPreview';
+export * from './ColorpickerDialog/StyledTooltipModal';
+export * from './ColorpickerDialog/StyledTooltipBody';
