--- packages/colorpickers/stories/2-ColorPicker.stories.tsx
@@ -1,17 +1,0 @@
-/**
- * Copyright Zendesk, Inc.
- *
- * Use of this source code is governed under the Apache License, Version 2.0
- * found at http://www.apache.org/licenses/LICENSE-2.0.
- */
-
-import { Meta } from '@storybook/react';
-import { ColorPicker } from '@zendeskgarden/react-colorpickers';
-
-export default {
-  title: 'Components/ColorPickers/ColorPicker',
-  component: ColorPicker
-} as Meta;
-
-export { Uncontrolled } from './examples/ColorPicker/Uncontrolled';
-export { Controlled } from './examples/ColorPicker/Controlled';
